# Conflicts:
#   frontend/src/api/drf.js
#   frontend/src/components/MainPage/RecentProblem.vue
#   frontend/src/components/ProblemModal/ProblemModal.vue

+++ prosn/src/main/java/com/ssafy/prosn/controller/StudyController.java
@@ -1,20 +1,29 @@
 package com.ssafy.prosn.controller;
 
 import com.ssafy.prosn.domain.study.StudyGroup;
+
 import com.ssafy.prosn.dto.*;
 import com.ssafy.prosn.repository.study.StudyGroupRepository;
 import com.ssafy.prosn.repository.study.UserStudyRepository;
 import com.ssafy.prosn.service.StudyService;
 import com.ssafy.prosn.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.Map;
+
+import static org.springframework.http.HttpStatus.*;
+
 
 /**
  * created by yeomyeong on 2022/08/01
- * updated by yeomyeong on 2022/08/02
+ * updated by yeomyeong on 2022/08/08
  */
 
 @RestController
@@ -26,58 +35,72 @@ public class StudyController {
     private final StudyService studyService;
     private final StudyGroupRepository studyGroupRepository;
     private final UserStudyRepository userStudyRepository;
+
     //스터디 그룹
     //생성
     @PostMapping
     public ResponseEntity<?> createStudy(@RequestBody @Valid StudyGroupRequestDto dto) {
         UserResponseDto user = userService.getMyInfoBySecret();
-        return ResponseEntity.ok(studyService.create(dto, user.getId()));
+        return ResponseEntity.status(OK).body(studyService.create(dto, user.getId()));
     }
+
     //목록 조회
     @GetMapping
-    public ResponseEntity<?> studyGroupList() {
-        return ResponseEntity.ok(studyGroupRepository.showStudyGroupList());
+    public ResponseEntity<?> studyGroupList(@PageableDefault(size = 5) Pageable pageable) {
+        Page<StudyGroupListResponseDto> page = studyGroupRepository.showStudyGroupList(pageable);
+        return ResponseEntity.status(OK).body(page);
     }
+
     //상세내용
     @GetMapping("/{studyid}")
     public ResponseEntity<?> showStudy(@PathVariable(value = "studyid") Long studyid) {
         UserResponseDto user = userService.getMyInfoBySecret();
-        return ResponseEntity.ok(studyService.showStudyGroup(user.getId(), studyid));
+        return ResponseEntity.status(OK).body(studyService.showStudyGroup(user.getId(), studyid));
     }
+
     //수정
     @PutMapping
     public ResponseEntity<?> updateStudy(@RequestBody @Valid StudyGroupRequestDto dto) {
         StudyGroup studyGroup = studyGroupRepository.findById(dto.getId()).orElseThrow(() -> new IllegalStateException("존재하지 않는 스터디입니다"));
-        return ResponseEntity.ok(studyService.update(studyGroup.getId(), dto));
+        if (!studyGroup.getUser().getId().equals(userService.getMyInfoBySecret().getId())) {
+            throw new IllegalStateException("스터디를 수정할 권한이 없습니다.");
+        }
+        return ResponseEntity.status(OK).body(studyService.update(studyGroup.getId(), dto));
     }
+
     //삭제
     @DeleteMapping("/{studyid}")
     public ResponseEntity<?> deleteStudy(@PathVariable(value = "studyid") Long studyid) {
         StudyGroup studyGroup = studyGroupRepository.findById(studyid).orElseThrow(() -> new IllegalStateException("존재하지 않는 스터디입니다"));
+        if (!studyGroup.getUser().getId().equals(userService.getMyInfoBySecret().getId())) {
+            throw new IllegalStateException("스터디를 삭제할 권한이 없습니다.");
+        }
         studyService.deleteStudy(studyGroup);
-        return ResponseEntity.ok("해당 스터디 삭제 완료");
+        return ResponseEntity.status(OK).build();
     }
 
     //내스터디
     //목록 조회
     @GetMapping("/me")
-    public ResponseEntity<?>  myStudyGroupList() {
+    public ResponseEntity<?> myStudyGroupList(@PageableDefault(size = 5) Pageable pageable) {
         UserResponseDto user = userService.getMyInfoBySecret();
-        return ResponseEntity.ok(userStudyRepository.findByUserId(user.getId()));
+        return ResponseEntity.status(OK).body(userStudyRepository.findByUserId(user.getId(), pageable));
     }
+
     //가입
     @PostMapping("/me")
-    public ResponseEntity<?> joinStudy(@RequestBody Long studyid) {
+    public ResponseEntity<?> joinStudy(@RequestBody Map<String, String> req) {
         UserResponseDto user = userService.getMyInfoBySecret();
-        StudyGroup studyGroup = studyGroupRepository.findById(studyid).orElseThrow(() -> new IllegalStateException("존재하지 않는 스터디입니다"));
-        return ResponseEntity.ok(studyService.joinStudy(user.getId(), studyGroup));
+        StudyGroup studyGroup = studyGroupRepository.findById(Long.parseLong(req.get("id"))).orElseThrow(() -> new IllegalStateException("존재하지 않는 스터디입니다"));
+        return ResponseEntity.status(CREATED).build();
     }
+
     //탈퇴
     @DeleteMapping("me")
-    public ResponseEntity<?> removeStudy(@RequestBody Long studyid) {
+    public ResponseEntity<?> removeStudy(@RequestBody Map<String, String> req) {
         UserResponseDto user = userService.getMyInfoBySecret();
-        StudyGroup studyGroup = studyGroupRepository.findById(studyid).orElseThrow(() -> new IllegalStateException("존재하지 않는 스터디입니다"));
+        StudyGroup studyGroup = studyGroupRepository.findById(Long.parseLong(req.get("id"))).orElseThrow(() -> new IllegalStateException("존재하지 않는 스터디입니다"));
         studyService.removeStudy(user.getId(), studyGroup);
-        return ResponseEntity.ok("해당 스터디 탈퇴 완료");
+        return ResponseEntity.status(OK).build();
     }
 }
